Create README.md
This week we look at Lists and Tuples. These two are closely related; initially it may seem that a Tuple is just a special type of List. The key things to look at are how these two are used, and how the (seemingly small) differences affect where they would be useful. You could, for example, always use a List in place of a Tuple.